Refractoring the Article route
