+++ server/.env
@@ -2,3 +2,4 @@ SUPERSEC= password123
 CLOUD_NAME=e-commerce-website
 CLOUD_API_KEY=742572971787234
 CLOUD_API_SECRET=4ytWgJj4DLzBDJNh2_J1kL2mpUU
+NODE_ICU_DATA="$(pwd)/node_modules/full-icu"
