+++ server/controller/articleController.js
@@ -2,7 +2,13 @@ const moment = require('moment');
 const pool = require('../models/db');
 const queries = require('../queries/articleQuery');
 require('../models/articleModel')();
+require("dotenv").config();
+const { DateTime } = require("luxon");
 
+const getTime = () => {
+	const easternTime = DateTime.local();
+	return easternTime.toLocaleString(DateTime.DATETIME_FULL);
+};
 
 class ArticleController {
   static async createArticle(req, res) {
@@ -43,6 +49,41 @@ class ArticleController {
       res.status(500).send({
         message:'Server Error',
         error: err.message
+      })
+    }
+  }
+
+  static async flagArticles(req,res){
+       try {
+            const {articleId} = req.params
+            const {flagged} = req.body
+
+            const article = await pool.query(queries.selectArticle, [articleId]);
+
+            if (article.rows.length === 0) {
+        return res.status(404).json({
+            status: 'Failed',
+            message:  'Article with this id not found',
+        });
+        }
+            if(!flagged){
+                return res.status(404).json({
+                    status: 'Error',
+                    message: 'Not provided'
+                })
+            }
+            const flaggedAt = getTime();
+            const flaged = await pool.query(queries.flag,  [flagged, flaggedAt, articleId])
+            return res.status(200).json({
+                status: 'success',
+                message: 'This Article is considered inappropriate for our community. So it has been flagged by some of the users here on this platform. We would require you to delete it within the next 24hrs',
+                flagged: flaged.rows[0].flagged,
+                FlaggedAt : flaggedAt
+            })
+       } catch (err) {
+          res.status(500).send({
+          message:'Server Error',
+          error: err.message
       })
     }
   }
