+++ server/queries/articleQuery.js
@@ -2,8 +2,14 @@ const createNewArticle = `INSERT INTO articles (title, article, flagged, created
         VALUES ($1, $2, $3, $4, $5, $6, $7) RETURNING *`;
 const selectCategory = `SELECT * FROM categories WHERE category_id =$1`;
 
+// FLAG QUERY
+const selectArticle = 'SELECT * FROM articles WHERE article_id =$1';
+const flag = 'UPDATE articles SET flagged =$1, flagged_at =$2 WHERE article_id = $3 RETURNING *'
+
 
 module.exports = {
   createNewArticle,
-  selectCategory
+  selectCategory,
+  selectArticle,
+  flag
 };
