+++ server/routes/articleRoute.js
@@ -6,4 +6,6 @@ const validation = require("../middleware/validation");
 
 router.post('/', authorization, validation.article, controller.createArticle);
 
+
+router.patch('/:articleId/flag',   controller.flagArticles);
 module.exports = router;
