--- server/routes/flaggedRoute.js
@@ -1,6 +1,0 @@
-const router = require('express').Router();
-const controller = require('../controller/flaggedController');
-
-router.patch('/:articleId',   controller.flagArticles);
-
-module.exports = router;
